% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bladder1_stacked.R
\docType{data}
\name{bladder1_stacked}
\alias{bladder1_stacked}
\title{Stacked version of the bladder1 dataset included in the survival package}
\format{
A data frame with 412 rows and 10 columns
}
\usage{
bladder1_stacked
data(bladder1_stacked)
}
\description{
This dataset includes the bladder1 dataset included the survival package,
which has been transformed into stacked format for use with \code{JointFPM}. The
stacked datset includes one row per individual for the competing event and
one rows per individual for each reoccurrence of bladder cancer.
}
\details{
For more information please take a look at \code{?survival::bladder}.
}
\keyword{datasets}
