% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointFPM-package.R
\docType{package}
\name{JointFPM-package}
\alias{JointFPM-package}
\title{JointFPM: A Parametric Model for Estimating the Mean Number of Events}
\description{
Implementation of a parametric joint model for modelling recurrent and competing event processes using generalised survival models. The joint model can subsequently be used to predict the mean number of events in the presence of competing risks at different time points. Comparisons of the mean number of event functions, e.g. the differences in mean number of events between two exposure groups, are also available.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/entjos/JointFPM}
  \item \url{https://entjos.github.io/JointFPM/}
  \item Report bugs at \url{https://github.com/entjos/JointFPM/issues}
}

}
\author{
\strong{Maintainer}: Joshua P. Entrop \email{joshuaentrop@posteo.de} (\href{https://orcid.org/0000-0003-1614-8096}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Alessandro Gasparini [contributor]
  \item Mark Clements [contributor]
}

}
\keyword{internal}
