% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_no.R
\name{mean_no}
\alias{mean_no}
\title{Non-parametric estimation of mean number of events}
\usage{
mean_no(formula, re_indicator, ce_indicator, data)
}
\arguments{
\item{formula}{A formula passed to `survfit`. The formula needs to be of form 
`Surv() ~ 1`, where `Surv()` need to be of `type == 'counting'`.}

\item{re_indicator}{The name of a variable indicating that these rows in the dataset belongs
to the risksets for the recurrent event.}

\item{ce_indicator}{The name of a variable indicating that these rows in the datasets belongs
to the riskset for the competing event.}

\item{data}{A `data.frame` in stacked format. The dataset needs to include one row
for the competing event and one row for each risk episode of the
recurrent event}
}
\description{
Non-parametric estimation of mean number of events
}
