% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.JointFPM.R
\name{predict.JointFPM}
\alias{predict.JointFPM}
\title{Post-estimation function for JointFPMs}
\usage{
\method{predict}{JointFPM}(JointFPM, type = "mean_no", newdata, t, nodes_start = 50, max_iter = 5)
}
\arguments{
\item{JointFPM}{A joint flexible parametric model of class \code{JointFPM}.}

\item{type}{A character vector defining the estimant of interest. Currently the only
option for this is "mean_no" for the mean number of events.}

\item{newdata}{A \code{data.frame} including the variable parameters for the prediction. One
value for each variable used in either the recurrent or competing event
model is requierd}

\item{t}{A vector defining the time points for the prediction.}

\item{nodes_start}{Number of nodes used for the initial Gaussian quadrature approximation
of the integral (default = 50):
\deqn{E[N(t)] = \int_{0}^{t} S(t)\lambda(t)}}

\item{max_iter}{The maximum number of iterations for the Gaussian quadrature
(default = 5).}
}
\value{
A \code{data.frame} with the following columns:
\itemize{
\item{\code{t}: }{The time for the prediction,}
\item{\code{fit}: }{The point estimate of the prediction,}
\item{\code{lci}: }{The lower confidence interval limit,}
\item{\code{uci}: }{The upper confidence interval limit.}
}
}
\description{
Predicts different estimates from a joint flexible parametric model.
Currently only the estimation of the mean number of events at different
time points is supported.
}
