% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_dfs.JointFPM.R
\name{test_dfs.JointFPM}
\alias{test_dfs.JointFPM}
\title{Tests DFs for FPMs}
\usage{
test_dfs.JointFPM(
  surv,
  re_terms,
  ce_terms,
  re_indicator,
  ce_indicator,
  df_ce,
  tvc_re,
  tvc_re_terms = NULL,
  tvc_ce_terms = NULL,
  cluster,
  data,
  same_dfs_tvc = FALSE,
  by_vars = NULL
)
}
\arguments{
\item{data}{A data set that should be used to fit the FPM.}

\item{same_dfs_tvc}{If `TRUE` no combinations of DFs between the different tvc variables will
be tested. Instead only DFs from the minimum DF specified until the
largest specified DF in `dfs_tvc` will be tested for all variables
specified in `dfs_tvc` at the same time.}

\item{by_vars}{A character vector of factor variable names used to fit stratified FPMs.
One FPM will be fitted for each combination of factor levels specified.
This is especially useful, when testing stratified FPMs.}

\item{formula}{A R-formula that is passed to the stpm2 call.}

\item{dfs_bh}{A vector of DFs to be tested for the baseline hazard function.}

\item{dfs_tvc}{A list of variables with a vector of DFs that should be tested for
time varying effects, e.g. `list(case = 1:10, female = 1:3)`.}
}
\value{
A `data.frame` with one row per combination of baseline hazards
   and tvc dfs and the corresponding AIC and BIC. If the `by_vars`
   argument is specified a `list` of `data.frame`s with one `data.frame`
   for each strata will be returned.
}
\description{
Test of degrees of freedom (DFs) for Flexible Parametric Survival Models
(FPMs).
}
\examples{
require(rstpm2)

data(brcancer)
data(colon)

fpm_test_dfs(Surv(rectime, censrec) ~ hormon + x1,
   dfs_bh  = 1:5,
   dfs_tvc = list(hormon = 1:3,
                  x1     = 1:5),
   data = brcancer)

fpm_test_dfs(Surv(rectime, censrec) ~ hormon + x1,
   dfs_bh  = 1:5,
   dfs_tvc = list(hormon = 1:3,
                  x1     = 1:5),
   same_dfs_tvc = TRUE,
   data = brcancer)

fpm_test_dfs(Surv(surv_mm, status == "Dead: cancer") ~ age + year8594,
             dfs_bh  = 1:5,
             dfs_tvc = list(age      = 1:5,
                            year8594 = 1:5),
             by_vars = c("sex", "stage"),
             same_dfs_tvc = TRUE,
             data = colon)
}
