% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointFPM-package.R
\docType{package}
\name{JointFPM-package}
\alias{JointFPM-package}
\alias{_PACKAGE}
\title{JointFPM: A Parametric Model for Estimating the Mean Number of Events}
\description{
A parametric model for estimating the mean number of events in the presence of competing events. The estimator is based on a joint flexible parametric model (FPM) for the recurrent and competing event process and includes a wrapper for 'rstpm2::stpm2()' which implements an easy specification of such models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/entjos/JointFPM}
  \item Report bugs at \url{https://github.com/entjos/JointFPM/issues}
}

}
\author{
\strong{Maintainer}: Joshua P. Entrop \email{joshuaentrop@posteo.de} (\href{https://orcid.org/0000-0003-1614-8096}{ORCID})

Other contributors:
\itemize{
  \item Alessandro Gasparini [contributor]
  \item Mark Clements [contributor]
}

}
\keyword{internal}
