% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.JointFPM.R
\name{summary.JointFPM}
\alias{summary.JointFPM}
\title{Summarises a \code{JointFPM} objects}
\usage{
\method{summary}{JointFPM}(object, ...)
}
\arguments{
\item{...}{Other arguments that should be passed to print.}

\item{x}{An \code{JointFPM} object.}
}
\value{
No return value, called for side effects.
}
\description{
This is a summary function for \code{JointFPM} objects, created with
\code{JointFPM()}. The function improves the readability of the output.
}
